### 2025-01-13 #7660 @editorialbot
[PRE REVIEW]: Credit Default Swaps with R
**Submitting author:** @yanrong-stacy-song (<a href="http://orcid.org/0009-0003-5738-8384">Yanrong Song</a>)
**Repository:** https://github.com/yanrong-stacy-song/creditr
**Branch with paper.md** (empty if default branch): main
**Version:** v0.6.2
