### 2025-01-17 #7660 @lrnv
@yanrong-stacy-song Greetings! I'll be your reviewer on this submission. Would you by chance know a few individuals that might be willing to review this work for JOSS ? We expect people that can understand the finance **and** know the standards of R/C implementations.
