### 2025-01-17 #7660 @yanrong-stacy-song
@lrnv Hi Oskar, nice to meet you and thanks for helping! Previously Joshua Ulrich helped review one of our finance R packages in 2018. Our co-author has briefly spoken to Dirk Eddelbuettel at previous R/Finance conferences but not sure he reviews now. Unfortunately I don't know someone else. Thx!
