### 2025-01-18 #7660 @lrnv
@yanrong-stacy-song I do not understand clearly what in the repository is your work or not. Basically all the C files have 2009 copyrights, so I guess that your contribution is only the R wrapper ? Could you please clarify exactly which pieces of code are yours are which arent ? It is not a problem in itself, but to correctly review your work we need to understand what exactly you are submitting.
