### 2025-01-18 #7660 @lrnv
@yanrong-stacy-song Also, could you please clarify exactly what are the contributions of each authors ? Since only you commited to the repository, it isnt clear to me that 8 people should share authorship. See [there](https://joss.readthedocs.io/en/latest/submitting.html#authorship) for authorship guidelines.
