### 2025-01-18 #7660 @yanrong-stacy-song
> [@yanrong-stacy-song](https://github.com/yanrong-stacy-song) I do not understand clearly what in the repository is your work or not. Basically all the C files have 2009 copyrights, so I guess that your contribution is only the R wrapper ? Could you please clarify exactly which pieces of code are yours are which arent ? It is not a problem in itself, but to correctly review your work we need to understand what exactly you are submitting.

@lrnv

- All files under the /src folder are all taken from the [ISDA standard CDS model](https://www.cdsmodel.com/), and we cited it in the DESCRIPTION and LICENSE files.
- Our contribution is the R wrapper (stuff not in the /src folder)
