### 2025-01-18 #7660 @yanrong-stacy-song
> [@yanrong-stacy-song](https://github.com/yanrong-stacy-song) Also, could you please clarify exactly what are the contributions of each authors ? Since only you commited to the repository, it isnt clear to me that 8 people should share authorship. See [there](https://joss.readthedocs.io/en/latest/submitting.html#authorship) for authorship guidelines.

@lrnv

The project was hosted by David Kane before he transferred the maintainership to me, and you can see all author contribution at these two places:

- https://github.com/yanrong-stacy-song/creditr/graphs/contributors (mostly mine)
- https://github.com/davidkane9/creditr/graphs/contributors (including others)

Yuanchu's contribution may seem missing from the Insights page, but I guess it's because his login / email is not sync'ed up, and you click into each commit, you will see his contribution
