### 2025-01-18 #7660 @lrnv
Much clearer about the scope of the submission. I guess that the paper could reflect a little more that this is a wrapper of another existing C library.

For the contributions, then, why didnt you fork the repository instead of initiating a new one ? Not necessarily a problem, just curious.
