### 2025-01-18 #7660 @yanrong-stacy-song
> Much clearer about the scope of the submission. I guess that the paper could reflect a little more that this is a wrapper of another existing C library.
>
> For the contributions, then, why didnt you fork the repository instead of initiating a new one ? Not necessarily a problem, just curious.

@lrnv We just thought it might look cleaner this way when submitting, but that may be unwarranted indeed. We are happy to merge back or turn one of the repos private too, if necessary
