### 2025-01-23 #7660 @lrnv
@yanrong-stacy-song Would merging back so that the main repository of the package has the full history and the full author list take a lot of effort from you ? If no, then yes that would be better for readability and comunication of authorship.

Then you can resubmit, ping me and I'll gladly take up the editing of the review based on the discussion we already had.

Otherwise, if you feel like its too much work, we could continue from here.
