### 2025-01-23 #7660 @lrnv
@joshuaulrich, @mitchellpound, @martinju, @n8thangreen, @coatless greetings! Would any of you be willing to review this submission for JOSS ? For those who do not know, JOSS is a journal that reviews code itself, its documentation can be browsed here : https://joss.readthedocs.io/en/latest/index.html, or you can simply ask me for any concerns you may have.
