### 2025-01-25 #7660 @yanrong-stacy-song
@lrnv Actually Andy Yao is a finance practitioner and previous published a paper at https://joss.theoj.org/papers/10.21105/joss.00061. Do you think we could ask him to review too? Thx!
