### 2025-01-28 #7660 @lrnv
@andyyao95, @n8thangreen, @ph-rast greetings! Would any of you agree to review this submission for JOSS ?
