### 2025-02-03 #7660 @lrnv
@joshuaulrich, @JustinMShea greetings! Would any of you agree to review this submission for JOSS ? if you do not already know, JOSS is a journal that reviews code itself, its documentation can be browsed here : [joss.readthedocs.io/en/latest/index.html](https://joss.readthedocs.io/en/latest/index.html), or you can simply ask me for any concerns you may have.
