### 2025-02-03 #7660 @expectopatronum
> [@andyyao95](https://github.com/andyyao95), [@n8thangreen](https://github.com/n8thangreen), [@ph-rast](https://github.com/ph-rast) [@expectopatronum](https://github.com/expectopatronum), greetings! Would any of you agree to review this submission for JOSS ?

I am sorry about the late response, but unfortunately I have to decline, I haven't been working with R in a couple of years.
