### 2025-02-07 #7758 @editorialbot
[REVIEW]: Credit Default Swaps with R
**Submitting author:** @yanrong-stacy-song (<a href="http://orcid.org/0009-0003-5738-8384">Yanrong Song</a>)
**Repository:** https://github.com/yanrong-stacy-song/creditr
**Branch with paper.md** (empty if default branch): main
**Version:** v0.6.2
