### 2025-02-08 #7758 @andyyao95
hi @lrnv. nice to meet you. this is my first time reviewing for the JOSS. do you mind providing some more details / tips? I've read through the guidelines.
