### 2025-02-10 #7758 @martinju
The license of the software seems to be non-standard and is thus not on the OSI-approved list: https://opensource.org/licenses/alphabetical
Do you have to use this license @yanrong-stacy-song, or would it be possible to switch to a standard OSI approved license?

Licensing is far from my expertise, but it seems to me that if it conflicts with the definition here. https://opensource.org/osd it may be a problem for the software? Can you clarify the requirements, @lrnv?
