### 2025-02-11 #7758 @yanrong-stacy-song
@martinju Hi Martin, thanks for reviewing! ISDA requires that we use this specific license (https://www.cdsmodel.com/cds-disclaimer.html?). The license text states:

> Licensor hereby grants You a world-wide, royalty-free, non-exclusive sublicensable license, subject to third party intellectual property claims,
a) to reproduce the ISDA CDS Standard Model in copies;
b) to distribute and publicly display the ISDA CDS Standard Model;
c) to use and authorize the use of the ISDA CDS Standard Model, with the proviso that copies of ISDA CDS Standard Model that you distribute for use by others shall be licensed under this License;
d) to translate, adapt, alter, transform, modify, or arrange the ISDA CDS Standard Model, thereby creating derivative works ("Derivative Works") based upon the ISDA CDS Standard Model...

So I think ISDA license does not conflict with https://opensource.org/osd. Thx!
