### 2025-02-11 #7758 @martinju
> ISDA requires that we use this specific license (https://www.cdsmodel.com/cds-disclaimer.html?)

I understand.

> So I think ISDA license does not conflict with https://opensource.org/osd. Thx!

You may very well be right that it is not conflicting with OSD (again not my expertise).
However, the license is not part of the [OSI approved](https://opensource.org/licenses/alphabetical) which is a point in our review checklist. It is not 100% clear for me whether this is a strict requirement or not for a JOSS submission.
https://joss.readthedocs.io/en/latest/submitting.html# says both: "The software must be open source as per the [OSI definition](https://opensource.org/osd)" which might be fine(?), but it also says "Make your software available in an open repository (GitHub, Bitbucket, etc.) and include an [OSI approved open source license](https://opensource.org/licenses)." which does not seem to be the case.

Thus, before I continue my review, I would like a clarification on the policies for such from @lrnv or other JOSS editors.

@lrnv Would you please take a look at this potential issue?
