### 2025-02-12 #7758 @yanrong-stacy-song
@lrnv Understood, what's the preferred way of fixing this? i.e. I can include a GPL-3 license in Github and point out the rest of the software (beyond original ISDA C code) is covered by GPL-3, and is that sufficient? Or do I need to do more? Thx!
