### 2025-02-13 #7758 @yanrong-stacy-song
@lrnv @martinju @andyyao95 I've updated the LICENSE file in my repo. Could you check and confirm it looks good now? Thx!
