### 2025-02-13 #7758 @lrnv
@yanrong-stacy-song  The editorialbot does not recognise your licence as an OSI-approved one, so you must have made a mistake somewhere. Can you check out the licencing part of the JOSS docs ?
