### 2025-02-13 #7758 @yanrong-stacy-song
@lrnv @martinju @andyyao95 Thanks for pointing that out. I've updated the LICENSE file again and now the editorial bot can recognize it as GPL-3. Could you confirm if it now looks good, and proceed if so? Thx!
