### 2025-02-22 #7758 @andyyao95
@yanrong-stacy-song i seem to have issue creating a CDS for more recent years. See example below. Is this intended behavior given data constraints?

cds2 <- CDS(date=as.Date("2020-03-14"),
            tenor=10,
            spread=100,
            currency='USD')

results in the following error:

Error in rates.URL(ratesURL) :
  Either an internet connection problem occurs or the data
        of the required date are unavailable on the internet. Try
        again later.
Error in x[, -1] : incorrect number of dimensions

but it works if i replace 2020 with 2015
