### 2025-02-24 #7758 @yanrong-stacy-song
@andyyao95 Good catch, after S&P/IHS/Markit mergers+acquisitions, the URL and XML changed a bit, and only limited history (rolling ~42 days) are provided online now. I have updated my Github repo (https://github.com/yanrong-stacy-song/creditr) to fix this.

You can either continue testing with pre-2015 dates or run `devtools::install_github` (I will submit to CRAN later too) to test with most recent dates . See [here](https://github.com/yanrong-stacy-song/creditr/blob/main/R/get_raw_markit.R#L82) for the comprehensive error string
