### 2025-03-21 #7758 @martinju
A few comment:

- [ ] There is a comprehensive vignette in the repo: https://github.com/yanrong-stacy-song/creditr/blob/main/vignettes/creditr.pdf The JOSS submission paper https://github.com/openjournals/joss-papers/blob/joss.07758/joss.07758/10.21105.joss.07758.pdf seems to be a copy of the introduction of this paper. That is not sufficient. The JOSS submission "paper" should itself describe the software. Please fix this. (Moreover, I see the CRAN version of the vignette has the JSS-affiliations on it -- please remove that in your next CRAN submission)
- [ ] There repository readme does not properly document the software. It should provide examples, installation instructions etc as described here: https://joss.readthedocs.io/en/latest/review_criteria.html#documentation (from statement of need all the way to Community guidelines)
