### 2025-05-08 #7758 @martinju
Hi again. I have now finished my review, and ticked of most of the bullets in the review checklist. While checking various aspects of the submission, I have identified the following issues. Please fix them and let me know when it is not done. If you disagree, have comments, need clarifications etc, just let me know in a comment.

### Code and dependencies
- [ ] adj_next_bus_day() - as I understand it, this function does not obey legal holidays. This is probably fine, but should be clearly stated. The "following 5D bus day convention" may indicate this, but I still suggest making it clearer.

- [ ] Quite a few typos in the documentation (e.g. build_rates() Details: "since Markit weibsite" -> "Since the Markit website", "over 1Y.So we" -> over 1Y. So we".
Please READ carefully go through all documentation and correct typos. You may also use devtools::spell_check to catch the most obvious ones.

- [ ] Do you need to expose and export all your functions to the users? Examples of what seems to be mainly internal functions not intended for users are call_ISDA, get_raw_markit and I am sure there are more.

- [ ] Do you really NEED all your imported packages? May some of them be listed under suggests instead? In particular, I find it odd that you need to import devtools.
In your creditr-package.R file you are importing all functions from all of these packages. Since you don't use :: when using functions from these packages, it is also impossible to find out which functions are actually used. I believe this is bad practice as opposed to importing only those functions you need. Functions from packages which are only needed for non-core functionality should be listed unders suggests instead and not imported in this way.
See https://r-pkgs.org/dependencies-mindset-background.html and https://r-pkgs.org/dependencies-in-practice.html for more details about this, where to put a package dependency and the potential issues related to importing all functions from lots of packages.
It is a bit of work to fix this, but I believe it is worth it.

- [ ] I see that you are currently failing on the additional "rchk" check on CRAN: https://cran.r-project.org/web/checks/check_results_creditr.html,
related to some unprotected variables in your C code. Would you be able to fix that, or is it not possible for you to fix issues with the cpp-code since I understand that is not your code? Please inform.

### Tests and examples

- [ ] Add local tests for the above download functions. Tests requiring internet access may not be suitable for CRAN, but since you need them to work and they have changed before, I suggest you add tests for them that you can run locally. Add them as regular testthat tests, but use the skip_on_cran() function to, well, skip them on cran.

- [ ] In your tests, you are using context() which is a superseded function from the testthat package. Please remove.
- [ ] Some dontrun-examples do not run.

This goes for the download_FRED() and download_markit() functions, which also affect the build_rates().
Please fix and ensure that all your dontrun-examples actually work (using e.g. devtools::run_examples(run_dontrun = TRUE)) (even if it takes a long time locally). I guess this is somewhat related to the issues reported by the other review above, but when testing it now, the same error is thrown regardless of the time point being shown.

- [ ] Readme - installation: In your github readme you are using devtools::install_github(). I suggest changing this to remotes::install_github() as remotes is a more lightweight package (where devtools gets its function from)

- [ ] Readme - example: I suggest expanding slightly on the example in your readme, perhaps showing a few other functions from your package, and providing some (commented out) output from your package

- [ ] Readme - Community guidelines: You should have info on how to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support. Please see the github repo of a recen JOSS paper for how this may look.

- [ ] Software paper: This looks mostly good now, I think. However, it is not perfectly clear for me that this is mostly a wrapper around cpp-code that is not yours. You should make that clear. Also, please mention that under "about" in your readme.
