+++ R/CS10.R
@@ -13,7 +13,7 @@
 #'   points. It specifies the payment amount from the protection buyer to the 
 #'   seller on an annual basis.
 #' @param recovery.var character, column in x containing recovery rates. ISDA
-#'   model standard recovery rate asscumption is 0.4.
+#'   model standard recovery rate assumption is 0.4.
 #' @param notional.var character, column in x containing the amount of the 
 #'   underlying asset on which the payments are based.
 #' @param notional numeric, the notional amount for all pricing if there isn't a
