+++ man/upfront_to_spread.Rd
@@ -36,7 +36,7 @@ seller on an annual basis.}
 \item{maturity.var}{character, column in x containing maturity date.}
 
 \item{recovery.var}{character, column in x containing recovery rates. ISDA
-model standard recovery rate asscumption is 0.4.}
+model standard recovery rate assumption is 0.4.}
 
 \item{upfront.var}{is the character name of upfront column}
 
