### 2025-05-22 f1ff130
move devtools to suggests
