+++ man/creditr.Rd
@@ -8,7 +8,7 @@
 \description{
 \code{creditr} package prices credit default swaps
 (CDS). It enables CDS class object which has slots as name, contract, RED, 
-date, spread, maturity, teno, coupon, recovery, currency, notional, 
+date, spread, maturity, tenor, coupon, recovery, currency, notional, 
 principal, accrual, pd, price, upfront, spread.DV01, IR.DV01 and rec.risk.01,
 with S4 methods like update, show and summary. It also supports data frame 
 input and is able to provide convenient calculation of key CDS statistics 
@@ -22,7 +22,7 @@ information and accounting conventions related to CDS pricing. Finally,
 \code{get_rates} and \code{build_rates} facilitates direct fetching of 
 relevant interest rates from online sources. Thanks to ISDA Standard Model's 
 Open Source license, we are able to create this package for R users. You can 
-find the Open Source licence of ISDA Standard Model at 
+find the Open Source license of ISDA Standard Model at 
 "https://www.cdsmodel.com/cdsmodel/cds-disclaimer.html"
 }
 \seealso{
