+++ DESCRIPTION
@@ -42,7 +42,6 @@ Authors@R: c(
 Depends: R (>= 3.1.0)               
 Imports: utils,
              quantmod,
-             devtools,
              methods,
              Rcpp (>= 0.10.6),
              RCurl,
@@ -50,7 +49,7 @@ Imports: utils,
              zoo,
              xts
 LinkingTo: Rcpp
-Suggests: testthat
+Suggests: devtools, testthat
 License: file LICENSE
 URL: https://github.com/yanrong-stacy-song/creditr
 Description: Price credit default swaps using
