### 2025-05-22 ebb3734
wrapper note in paper and readme
