### 2025-05-22 4490ef1
Create CONTRIBUTING.md
