+++ tests/testthat/test_CS10.R
@@ -1,4 +1,4 @@
-context("Test CS10")
+## Test CS10
 
 test_that("test case for CS10", {
   x <- data.frame(date = c(as.Date("2014-04-22"), as.Date("2014-04-22")),
