### 2025-05-27 #7758 @yanrong-stacy-song
@martinju @lrnv Here are my first set of improvements.

- `adj_next_bus_day` documentation has been updated: https://github.com/yanrong-stacy-song/creditr/pull/1
- Doc typos have been fixed: https://github.com/yanrong-stacy-song/creditr/pull/2
- Where do you see `call_ISDA` and `get_raw_markit` exported? I actually don't see them exported
- Changed `devtools` from imports to suggests: https://github.com/yanrong-stacy-song/creditr/pull/3
- Specify imported functions with `::` and reduce package-level imports: https://github.com/yanrong-stacy-song/creditr/pull/4
- The CRAN `rchk` check failure is indeed related to ISDA's standard C code, not controlled by me
- Removed `testthat::context()`: https://github.com/yanrong-stacy-song/creditr/pull/8
- Changed `devtools::install_github` to `remots::install_github` in Readme: https://github.com/yanrong-stacy-song/creditr/pull/7
- Added community guidelines in Readme: https://github.com/yanrong-stacy-song/creditr/pull/6 and https://github.com/yanrong-stacy-song/creditr/pull/7
- Further differenciated between ISDA C code and our R code in Readme and JOSS paper: https://github.com/yanrong-stacy-song/creditr/pull/5
