### 2025-06-06 #7758 @martinju
Thanks for the work @yanrong-stacy-song
Some comments below:

> Where do you see call_ISDA and get_raw_markit exported? I actually don't see them exported

Your NAMESPACE has `exportPattern("^[[:alpha:]]+")` stemming from creidtr-package.R `#' @exportPattern "^[[:alpha:]]+"`, which essentially means you are exporting almost all your functions, I guess. I suggest you remove that, and rather use `@export` for the functions you really want to export. Moreover, your may use `@internal` for functions you for some reason want to export, but you don't want to generate and expose the documentation to the end user. Note also that you should not have examples for non-exported functions.

> Specify imported functions with :: and reduce package-level imports: https://github.com/yanrong-stacy-song/creditr/pull/4

Good. I still see some `@import` in your files.  I will highly recommend to remove those and replace them by @importFrom PKG FUNCTION, and put them in the creditr-package.R as described in https://r-pkgs.org/dependencies-in-practice.html#sec-dependencies-in-imports OR just use PKG::FUNCTION for the functions not used that often.

After these updates, make sure to run R CMD CHECK to check that you don't get errors or warnings.

The rest of your fixes are good I think, leaving us with 4 remaining bullet points, I think. Please let me know when you have fixed them, or if you have further questions :-)
