+++ man/add_conventions.Rd
@@ -30,6 +30,7 @@ conventions added to it.
 x <- data.frame(date = c(as.Date("2014-05-06"), as.Date("2014-05-07")),
                 currency = c("USD", "JPY"))
 add_conventions(x)
+
 }
 \references{
 \url{
