+++ man/CDS.Rd
@@ -53,4 +53,5 @@ Build a \code{CDS} class object given the input about a CDS contract.
 }
 \examples{
 x <- CDS(date = as.Date("2014-05-07"), tenor = 5, spread = 50, coupon = 100) 
+
 }
