+++ NAMESPACE
@@ -1,13 +1,18 @@
 # Generated by roxygen2: do not edit by hand
 
+export(CDS)
+export(CS10)
+export(IR_DV01)
+export(add_conventions)
+export(add_dates)
+export(check_inputs)
+export(get_rates)
+export(rec_risk_01)
+export(spread_DV01)
+export(spread_to_pd)
 exportClasses(CDS)
 exportMethods(show)
 exportMethods(summary)
-exportPattern("^[[:alpha:]]+")
-import(Rcpp)
-import(methods)
-import(xts)
-import(zoo)
 importFrom(utils,head)
 importFrom(utils,tail)
 importFrom(utils,unzip)
