### 2025-06-17 #7758 @yanrong-stacy-song
@martinju Thx for revieiwng. I have replaced the blanket export pattern with specific exports which are needed for users and the vignette. I have also removed blanket imports and replace with either importFrom or ::. The PR is https://github.com/yanrong-stacy-song/creditr/pull/9.

I will work on the remaining 4 bullet points now. Thx again!
