+++ man/download_markit.Rd
@@ -27,10 +27,9 @@ with the yields for different maturities.
 }
 \examples{
 \dontrun{
-download_markit(start = as.Date("2005-12-31"), end = as.Date("2006-01-04"), 
-                currency = "JPY")
+download_markit(start = Sys.Date() - 10, Sys.Date(), currency = "JPY")
 }       
 }
 \seealso{
-\code{\link{download_FRED}} \code{\link{build_rates}}
+\code{\link{build_rates}}
 }
