+++ DESCRIPTION
@@ -41,7 +41,6 @@ Authors@R: c(
   )
 Depends: R (>= 3.1.0)               
 Imports: utils,
-         quantmod,
          methods,
          RCurl,
          XML,
