### 2025-08-04 #7758 @martinju
> [@lrnv](https://github.com/lrnv) I've also fixed the dontrun examples in this PR ([yanrong-stacy-song/creditr#10](https://github.com/yanrong-stacy-song/creditr/pull/10)). Could you check if your item lists are all done now? [@martinju](https://github.com/martinju) Thx!

Thanks, I will take a look and let you know.
