### 2025-08-08 #7758 @martinju
1. There are still issues with the exampes. Running `devtools::run_examples(run_donttest = TRUE, run_dontrun = TRUE)`
throws the first error at the build_rates function which is no longer exported. You cannot (according to CRAN https://stackoverflow.com/questions/34196597/r-package-example-function-cant-be-tested-if-the-function-is-not-exported have examples of non-exported functions).

2. running devtools::test() shows there are 8 empty tests:
```
── Skipped tests (8) ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
• empty test (8): test_add_conventions.R:11:1, test_CS10.R:3:1, test_get_rates.R:3:1, test_IR_DV01.R:3:1, test_rec_risk_01.R:3:1,
  test_spread_DV01.R:5:1, test_spread_to_pd.R:3:1, test_upfront_to_spread.R:5:1

```
So they are not testing anything? Not sure what the intention is?
3.  R CMD check also gives this error, which you also need to fix.
```
checking whether the namespace can be loaded with stated dependencies ... NOTE

  Warning: no function found corresponding to methods exports from 'creditr' for: 'show'

  A namespace must be able to be loaded with just the base namespace
  loaded: otherwise if the namespace gets loaded by a saved object, the
  session will be unable to start.

  Probably some imports need to be declared in the NAMESPACE file.
```

Once these 3 issues are fixed, and

`devtools::run_examples(run_donttest = TRUE, run_dontrun = TRUE)`
`devtools::test()`
`devtools::check()`

**ALL** run without notes, warnings or errors (please check this yourself), you should be good to go :-)
