### 2025-08-08 b48ee2d
remove examples for unexported fns and fix empty tests
