### 2025-08-10 #7758 @yanrong-stacy-song
Hi @martinju , thx for replying! I've fixed your first two points in this PR (https://github.com/yanrong-stacy-song/creditr/pull/11). For your point on R CMD Check NOTE, I wasn't able to replicate it on my Mac. Which platform are you using?
