+++ .github/workflows/rhub.yaml
@@ -77,10 +77,21 @@ jobs:
 
     steps:
       - uses: r-hub/actions/checkout@v1
+
       - uses: r-hub/actions/setup-r@v1
         with:
           job-config: ${{ matrix.config.job-config }}
           token: ${{ secrets.RHUB_TOKEN }}
+
+      # --- Added: install LaTeX so Sweave (.Rnw) vignettes can build ---
+      - name: Install TinyTeX (LaTeX)
+        uses: r-lib/actions/setup-tinytex@v2
+
+      # Optional: a few commonly-needed LaTeX packages for vignettes
+      - name: TinyTeX: install common LaTeX packages (optional)
+        run: tlmgr install fancyvrb upquote inconsolata xcolor microtype
+      # -----------------------------------------------------------------
+
       - uses: r-hub/actions/platform-info@v1
         with:
           token: ${{ secrets.RHUB_TOKEN }}
