### 2025-08-15 4e214ff
add importFrom methods show
