### 2025-08-15 #7758 @yanrong-stacy-song
Hi @martinju , thanks for replying, yes the NOTE occurs on R 4.5. I've fixed it in this PR (https://github.com/yanrong-stacy-song/creditr/pull/13). Could you check again? Thx!
