### 2025-08-19 #7758 @martinju
I confirm that it passes the checks now. I recommend accepting the paper/package as is now.
