### 2025-08-25 #7758 @lrnv
@yanrong-stacy-song:

1) The archive title should match the paper's title
2) The archive author list should match the paper's

Can you modify the archive accordingly ?
