### 2025-08-27 e4d46a7
add missing doi
