+++ paper.bib
@@ -21,7 +21,8 @@ TITLE     = {{Exchange Rates and International Financial Economics: History, The
 PUBLISHER = {{Palgrave Macmillan}},
 YEAR      = {2013},
 Month     = {Oct.},
-Note      = {}
+Note      = {},
+doi       = {https://doi.org/10.1057/9781137318886}
 }
 
 @ARTICLE{rates,
