### 2025-08-27 #7758 @yanrong-stacy-song
@lrnv Thanks for double checking. Indeed, "Exchange Rates and International Financial Economics" actually has a DOI. I've verified that the others don't have a DOI. I've filled in the missing DOI in paper.bib and pushed to Github. Could you check again? Thx!
