### 2025-08-28 a4adc67
Update paper.bib
