+++ paper.bib
@@ -22,7 +22,7 @@ PUBLISHER = {{Palgrave Macmillan}},
 YEAR      = {2013},
 Month     = {Oct.},
 Note      = {},
-doi       = {https://doi.org/10.1057/9781137318886}
+doi       = {10.1057/9781137318886}
 }
 
 @ARTICLE{rates,
