### 2025-08-30 #7758 @yanrong-stacy-song
@crvernon @lrnv I've fixed the invalid DOI, could you try again now? Thx!
