### 2025-09-01 #7758 @crvernon
@yanrong-stacy-song - there are still no in-text citations and no links or DOI showing in your references section.

You can check your changes after you address these by commenting:
`@editorialbot generate pdf`
