+++ paper.bib
@@ -33,6 +33,14 @@ doi       = {10.1057/9781137318886}
    YEAR= {2013}
 }
 
+@misc{isdaCDS2022,
+  title        = {ISDA CDS Standard Model},
+  author       = {{S\&P Global Market Intelligence}},
+  howpublished = {\url{https://www.cdsmodel.com/}},
+  note         = {Last updated in 2022},
+  year         = {2022}
+}
+
 @BOOK{creditDerivatives,
 AUTHOR = {Chacko, George and Sjoman, Anders and Motohashi, Hideto and Dessain, Vincent},
 TITLE = {{Credit Derivatives: A Primer on Credit Risk, Modeling, and Instruments}}, 
