### 2025-09-05 #7758 @yanrong-stacy-song
@crvernon  Hi Chris, thank you for accepting the paper!! Could we keep this on hold for few days? Before publishing, we'll need notify our compliance department first, thank you!
